requirements.txt
langchain
# duckduckgo-search 
langchain-community
python-dotenv
